pyproject.toml
[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120

[tool.black]
line-length = 120

[tool.poetry]
name = "life-assist"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0.2"
psycopg2-binary = "2.9.9"
pillow = "10.2.0"
crispy-bootstrap4 = "2023.1"
django-crispy-forms = "2.1.0"
django-bootstrap-datepicker-plus = "^5.0.5"
django-view-breadcrumbs = "2.5.1"
django-extensions = "^3.2.3"
django-plotly-dash = "^2.3.1"
dpd-static-support = "^0.0.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
